final commit

+++ task03polymorphism/src/main/java/by/jwd/task03polymorphism/service/NetWeightComparator.java
@@ -8,6 +8,6 @@ public class NetWeightComparator implements Comparator<ItemOfCoffee> {
 
 	@Override
 	public int compare(ItemOfCoffee item1, ItemOfCoffee item2) {
-		return ((int) item1.getCoffee().getNetWeight() - (int) item2.getCoffee().getNetWeight());
+		return (item1.getCoffee().getNetWeight() - item2.getCoffee().getNetWeight());
 	}
 }
